+++ pyvo/dal/tests/test_sia.py
@@ -61,7 +61,9 @@ class TestSIAService:
     @pytest.mark.filterwarnings("ignore::astropy.io.votable.exceptions.W42")
     @pytest.mark.filterwarnings("ignore::astropy.io.votable.exceptions.W49")
     def test_search(self):
-        service = SIAService('http://example.com/sia')
+        service = SIAService('http://example.com/sia', capability_description="A SIA service.")
+        assert service.baseurl == 'http://example.com/sia'
+        assert service.capability_description == "A SIA service."
 
         results = service.search(pos=(288, 15))
         result = results[0]